Factored out StubRegistryClientFactory
